tidy: Fix "multiplication implicit widening" in mythfrontend.
Fix the clang-tidy "implicit widening of multiplication result"
warnings.  This check is looking for places where a multiplication
operation may have overflowed the smaller type before the result is
converted to the bigger type. This will give erroneous results.

This set of warnings are false positives as the compiler is generating
constant values that are well within the range of the type. Fix the
warnings by either changing one of the constants to a 'long long' or
by enlarging the type of a variable.

+++ mythtv/programs/mythfrontend/playbackbox.h
@@ -52,7 +52,7 @@ using ProgramMap = QMap<QString,ProgramList>;
 using Str2StrMap = QMap<QString,QString>;
 using PlaybackBoxCb = void (PlaybackBox::*)();
 
-static constexpr int kMaxJobs {7};
+static constexpr size_t kMaxJobs {7};
 
 enum {
     kArtworkFanTimeout    = 300,
